Flag Travis builds that contain syntax errors
Ready for review.  This is a "rebase" of #108

Use flake8 to prevent pull requests that contain Python syntax errors, undefined names, or tab indentation.

For the past 9 months, we have been running flake8 in --exit-zero mode which treats all flake8 errors as warnings. This has helped us to find and fix many important errors. Our issue counts are now reasonable: Py2, Py3.

So, the time has come to select a handful of the the most serious errors and fail the Travis CI build if new PRs introduce those errors. We do this by making two flake8 runs in our Travis process. The first run can fail the build but only focuses on a few serious errors and the second runs in --exit-zero mode which will continue our current process of reporting but not failing the build on "style" violations.

Delete .flake8 to increase flexibility

We want to be able to make these choices at runtime