Fix an issue where the same port couldn't be used twice for an incoming tcp connection if the process that created the first was still active.
This commit is a one-line swap in order to set the socket option for address reusal before binding to the specified port. Originally this led to a bug where the port would refuse additional connections if the process that created the first connection was still active, even if the socket had been closed. By swapping these lines, the ports can be re-used. This fix was tested on OSX Mojave v10.14.4.