mavmmaplog: Close the mmap object
The inherited close() method from the mavlogfile class only closes the file object. However, the mmap object also keeps a handle on the file, which prevents it from being deleted or opened in read mode even after we call close() This commit closes both the file object (by calling the super method) and the mmap object (directly), ensuring that all handles are freed.