mavutil: return mode as 'Mode(msg.custom_mode)' for px4 HIGH_LATENCY2
Before, when reading HIGH_LATENCY2 messages that came from a PX4
controller mavutil threw an exception because after the check for PX4,
`msg.base_mode` was read immediately. It was not regarded that the
`HIGH_LATENCY2` message does not contain a `base_mode` field.

This introduces a fix where the mode string for HIGH_LATENCY2 messages
from a PX4 controller is returned as `Mode(msg.custom_mode)`.