mavlink_helpers.h: Don't restart parsing on second CRC byte
If an unknown message is received it would fail parsing due
to bad CRC, thanks to the crc_extra. Before this change the
parser would restart parsing on the second CRC byte. This
can cause the parser to be out of sync and drop messages if
the second CRC byte happens to be MAVLINK_STX.
Now the parser will ignore the MAVLINK_STX if it was
received as the second byte of a bad CRC message.