Mavgen WLUA: Fix incorrect display of negative ints on mac
Fix is to not include the value argument when calling add_le function, and let Wireshark
do the value extraction internally. This avoids an issue where a double being cast to
a unsigned int (which should have been an int), becomes 0 on mac, but works on Linux.

The value argument is still required in the case when a float is used to represent an
enum or bitmask.
Also, value is used seperately when adding unit conversions and when doing any bitmask
extraction