Update requirements.txt for serial connection
Although pymavlink supports the use of serial connections with a flight control board and utilizes a serial connection package, this information is not reflected in the requirements.txt file.

Simply running "pip install -r pymavlink/requirements.txt" and attempting to use the mavutil module over serial will not work as expected.