In Ada generator don't use enumeration types.
Let's use integer types of the appropriate size instead of the
enumerated type. This allows the use of a value not specified
in the enumeration, which often happens when extending the
protocol (for example, adding new commands). Using static
functions instead of contants makes it possible to overload
names and at the same time does not interfere with the work
of case operators and other constructions where static values
are needed. With overloadable names we can drop common prefix
and have shorten names.