DFReader: add strict option to skip_to_type
skip_to_type adds all these extra messages to the parsed types just in
case the user wants to check conditions on those messages. This is a
waste in many cases, but especially when recv_match has not been given a
conditions argument. We could actually try to parse the conditions
for message types, but this is easier for now.