DFReader: init_clock fix bug
'GPS' was being passed as a condition, and we were not rewinding after.
Because of the first bug, we were effectively just using the old init
logic, which was correct, but slow in logs with no fix.

Additionally, I was not breaking at the right time. This would have
only worked for logs with a good time in the first GPS message.