generator: Remove dead code sha256_test.py
Fixes: #1062
* #1062

This file defines no classes or functions.  It can only be run on the command line, but it will always fail on Python 3.

Given that it is only a command-line tool and has not been used on Python 3, let's remove it as dead code.

---

A TypeError will be raised if this script is run on Python 3 in its current form:
```python
>>> import hashlib
>>> hashlib.new('sha256').update(b'bytes')  # OK
>>> hashlib.new('sha256').update('str'). # Fails!
Traceback (most recent call last):
  File "<python-input-1>", line 1, in <module>
    hashlib.new('sha256').update('str')
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^
TypeError: Strings must be encoded before hashing
>>> hashlib.new('sha256').update('str'.encode('utf-8'))  # OK
```

The problem is that `sys.argv` is a list of `str`, but `hashlib.new('sha256').update()` requires `bytes`, not `str` in Python 3.

This is not a problem in Py2 (`bytes == str` is True), but is a problem in Py3 (`bytes == str` is False).