Revert "Mavgen C: Use strncpy instead of memcpy to pack char[n] fields"
This reverts commit ff57508f739f49bf09e8ef3f523f17a44478d6c5.

The change broke the param_value field of PARAM_EXT_... messages which
is defined as char[128] and can contain arbitrary data such as
int/float or binary data.

Ideally, the type would be defined as uint8_t[128] and the issue
avoided, however, this can't be changed at this stage as it would change
the CRC_EXTRA and make the message incompatible.

Unfortunately, I don't see another way to fix this than to revert this
change.