Added support for version 2 protocol generation for Ada.
Added MAVDeprecated class to mavparse.py

Added deprecated field to MAVType, MAVEnumEntry, MAVEnum classes that are filled from the XML specification to generate corresponding comments.

Added invalid field to MAVField class that is filled from XML specification and can be used in generation. For example, to set as an initial (non-initialized) value for the field.